+++ src/game/Guild.cpp
@@ -1530,8 +1530,8 @@ void Guild::LoadBankRightsFromDB(uint32 GuildId)
         Field *fields = result->Fetch();
         uint8 TabId = fields[0].GetUInt8();
         uint32 rankId = fields[1].GetUInt32();
-        uint16 right = fields[2].GetUInt16();
-        uint16 SlotPerDay = fields[3].GetUInt16();
+        uint16 right = fields[2].GetUInt32();
+        uint16 SlotPerDay = fields[3].GetUInt32();
 
         SetBankRightsAndSlots(rankId, TabId, right, SlotPerDay, false);
